nvif: split out ctxdma interface definitions
